pharmacy/requirements.txt
## Requirements
Python==3.11
black==24.4.2
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8